requirements.txt
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.9.2
matplotlib>=1.4.3
pandas>=0.19
click==7.0
pillow==6.2.0
pytesseract==0.2.6